requirements.txt
# ================================================================
# AI-AGENT v2.0 - REQUIREMENTS FINAUX (Conception Mise à Jour)
# Architecture: FastAPI + RabbitMQ + PostgreSQL + LangGraph + AI Hub
# ================================================================

# ================================================================
# 1. API WEB FRAMEWORK (FastAPI + Serveur)
# ================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
gunicorn==21.2.0

# ================================================================
# 2. CELERY + RABBITMQ (Background Processing)
# ================================================================
celery[redis]==5.3.4
kombu==5.3.4
# Redis retiré - remplacé par RabbitMQ
pika==1.3.2                    # Client RabbitMQ Python
amqp==5.2.0                    # AMQP protocol pour RabbitMQ

# Monitoring Celery
flower==2.0.1

# ================================================================
# 3. BASE DE DONNÉES (PostgreSQL uniquement)
# ================================================================
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0

# Vector Store (pgvector pour recherche sémantique)
pgvector==0.2.4

# ================================================================
# 4. LANGGRAPH + LANGCHAIN (Workflow AI)
# ================================================================
langgraph==0.2.14
langgraph-checkpoint==1.0.12

# LangChain Core
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-text-splitters==0.2.4

# LangChain Providers
langchain-openai==0.1.23
langchain-anthropic==0.1.23

# ================================================================
# 5. AI PROVIDERS (Multi-Provider Hub)
# ================================================================
anthropic==0.64.0
openai==1.102.0

# ================================================================
# 6. HTTP CLIENTS (API Calls)
# ================================================================
httpx==0.25.2
requests==2.31.0
aiohttp==3.12.15

# ================================================================
# 6.1. COMMUNICATION PLATFORMS (Slack, etc.)
# ================================================================
slack-sdk==3.27.0             # Slack API client pour notifications

# ================================================================
# 7. GIT + GITHUB (Code Management)
# ================================================================
PyGithub==2.1.1               # GitHub API (version stable)
gitpython==3.1.40

# ================================================================
# 8. VALIDATION + CONFIGURATION (Pydantic)
# ================================================================
pydantic>=2.5.0,<2.8.0
pydantic-settings>=2.1.0,<2.3.0
python-dotenv==1.0.0
environs==10.3.0

# ================================================================
# 9. SÉRIALISATION JSON
# ================================================================
orjson==3.10.7
jsonpatch==1.33

# ================================================================
# 10. TESTS + COVERAGE (Pytest Suite)
# ================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
coverage==7.3.4

# ================================================================
# 11. QUALITY ASSURANCE (Linters + Formatters)
# ================================================================
pylint==3.0.3
flake8==6.1.0
black==23.11.0
isort==5.12.0
bandit==1.7.5                 # Security scanner
mypy==1.8.0                   # Type checker
mypy-extensions==1.1.0

# ================================================================
# 12. SÉCURITÉ + AUTHENTIFICATION (OAuth Monday.com + JWT Admin)
# ================================================================
cryptography==43.0.3
pyjwt[crypto]==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# OAuth2 Monday.com
authlib==1.2.1
python-jose[cryptography]==3.3.0

# Validation email pour Pydantic
email-validator==2.3.0
dnspython==2.8.0

# ================================================================
# 13. LOGGING + MONITORING
# ================================================================
structlog==23.2.0
rich==13.7.0
colorlog==6.8.0

# Monitoring système (optionnel)
psutil==5.9.6

# ================================================================
# 14. OUTILS GÉNÉRAUX (Helpers)
# ================================================================
aiofiles==23.2.1
python-dateutil==2.8.2
click==8.2.1
pyyaml==6.0.1
toml==0.10.2

# ================================================================
# 14.1. DATA PROCESSING + EXCEL (pour Golden Datasets)
# ================================================================
pandas==2.1.4
openpyxl==3.1.2

# ================================================================
# 15. NLP TOOLS (pour analyze_requirements node)
# ================================================================
nltk==3.8.1
spacy==3.7.2

# ================================================================
# 16. DÉVELOPPEMENT + UTILS
# ================================================================
watchfiles==0.21.0

# ================================================================
# 17. STANDARDS PYTHON
# ================================================================
urllib3>=1.26.0,<3.0.0
certifi==2025.8.3

# Extensions de types
typing-extensions>=4.5.0

# ================================================================
# NOTES D'INSTALLATION:
# ================================================================
# 1. pip install -r requirements.txt
# 2. python -m spacy download fr_core_news_sm (pour NLP français)
# 3. python -m nltk.data download stopwords
# 4. python -m nltk.data download punkt
# ================================================================ aio-pika>=9.0.0
